Latent Space Bootstrap

Add explanation of how latent space bootstrap works

Usage

bootstrap_latent_space(network, d = 2, B = 1000, output.type = "igraph")

Arguments

network

An igraph object with \(n\) nodes.

d

Dimension of latent space. Recommended only 2 or 3 to maintain interpretability.

B

number of bootstrap samples to calculate

output.type

The class of object the resulting bootstrap networks should be. The default is igraph which will make bootstrap samples of class "igraph". Note that for large \(B\), this may not be an efficient use of storage space. Other options include matrix which will return bootstrap samples as an \(n\)-by-\(n\) adjacency matrix, dgCMatrix which will return scarce matrices (package Matrix must be loaded).

Value

A list of length \(B\) where each element is an bootstrap sample. Each element is of class output.type.

Details

Cite Sussman et al. (2012) paper for ASE CITE Levin and Levina (2022) paper for bootstrap procedure. Link to Matrix package

Examples

# bootstrap_latent_space(paul.igraph, d=2, B=10, output.type = "matrix")