Detect the Type of Objects in a List

Attempts to detect the type of network representation in a list. Networks can be represented as "igraph", "dgCMatrix", "edgelist", or "matrix".

Usage

detect_type(obj)

Arguments

obj

A list of networks. All networks in the list must have the same representation. e.g. all networks in the list must be an edge list.

Value

A character string indicating the representation of the networks in the list. Possible return values are:

"igraph": If the networks are igraph objects.

"dgCMatrix": If the networks are sparse matrices of class dgCMatrix.

"edgelist": If the networks are matrices with two columns, representing an edge list.

"matrix": If the networks are square matrices, typically representing adjacency matrices.

Details

The function detects if the input list of networks are represented as "igraph", "dgCMatrix", "edgelist" or "matrix". This function is primarily used internally for verifying the networks are in the correct format for the intended operation being used.

2-by-2 matrices default to "matrix".

See also

Examples

# Convert list of adjacency matrices to igraph objects
one_matrix <- matrix(c(0, 1, 1, 1, 0, 0, 1, 0, 0), nrow = 3, byrow = TRUE)
matrix_list <- replicate(3, one_matrix, simplify = FALSE)
igraph_list <-
  make_network_type(matrix_list,
                   input.type = "matrix",
                   output.type = "igraph")

# Same thing now using detect type
type <- detect_type(matrix_list)
type
#> [1] "matrix"
igraph_list <-
  make_network_type(matrix_list,
                   input.type = type,
                   output.type = "igraph")

# Convert list of igraph objects to sparse matrix
sparsematrix_list <-
  make_network_type(igraph_list,
                    input.type = "igraph",
                    output.type = "dgCMatrix")
detect_type(sparsematrix_list)
#> [1] "dgCMatrix"