Jackknife-after-Bootstrap for Linear Regression

ADD DESCRIPTION - add something that the default is 0.05-0.95 i.e. center 90%- why default for B =3100 ADD CITATIONS

Usage

jab_lm(
  mod,
  stat = "rstudent",
  quant.lower = 0.05,
  quant.upper = 0.95,
  B = 3100,
  package.name = NULL,
  stat.args = NULL
)

Arguments

mod

An lm object. Output from lmstats.

stat

A character string of the function name used to calculate the desired centrality statistic. The function must input an lm model object as is first argument and output a length \(n\) vector of the statistic of interest.

quant.lower

A numeric between 0 and 1 used as the lower cutoff in the JaB algorithm. Default is 0.05. Must be smaller than quant.upper.

quant.upper

A numeric between 0 and 1 used as the upper cutoff in the JaB algorithm. Default is 0.95. Must be larger than quant.lower.

B

Number of bootstrap samples. Default is 3100

package.name

A character string of the name of the package that func.name function is in. If left as NULL if the function will be called as loaded in the users environment.

stat.args

A named list of additional arguments the func.name function may need beyond the mod object.

Value

A data frame with 5 columns.

"row.ID": Row Number of the observation in the data set used for mod.

"lower": Lower quantile cutoff as determined by quant.lower.

"upper": Upper quantile cutoff as determined by quant.upper.

"orig": The original influence/outlier statistic calculated by stat.

"influential": Logical flagging if the observation is influential or not. TRUE if orig < lower or orig > upper. FALSE otherwise.

Examples

library(stats)
data("LifeCycleSavings")

mod <- lm(sr ~ ., data = LifeCycleSavings)

# JaB with DFFITS
result1 <- jab_lm(mod,
                  stat = "dffits",
                  quant.lower = 0.025,
                  quant.upper = 0.975,
                  B = 3100)
result1[result1$influential, ]
#>        row.ID      lower     upper       orig influential
#> Japan      23 -0.5962788 0.6242944  0.8596508        TRUE
#> Zambia     46 -0.6127221 0.6599902  0.7482351        TRUE
#> Libya      49 -0.5972119 0.6806704 -1.1601334        TRUE

#define the likelihood distance as influence statistic
## NEED TO CITE
infl_like <- function(mod){

  n <- length(mod$fitted.values)
  p <- length(mod$coefficients)

  ti <- rstudent(mod)
  h <- hatvalues(mod)

  p1 <- log( (n/(n-1)) * ((n-p-1) / (ti^2 +n-p-1)) )
  p2 <- ti^2 * (n-1) / (1-h) / (n-p-1)

  return(n*p1 + p2 - 1)
}

## JaB with Likelihood Distance
result2 <- jab_lm(mod,
                  stat = "infl_like",
                  quant.lower = 0.00,
                  quant.upper = 0.95,
                  B = 3100)
#> Error in jab_lm(mod, stat = "infl_like", quant.lower = 0, quant.upper = 0.95,     B = 3100): The functioninfl_likedoes not exist in your environment.
result2[result2$influential, ]
#> Error in eval(expr, envir, enclos): object 'result2' not found