Sample about one seed from snowboot package

This is a slightly modified version of the sample_about_one_seed from the snowboot package.

Usage

sample_about_one_seed_modified(net, seed, n.wave = 1)

Arguments

net

A network object as described in the snowboot package. package

seed

Character string of the selected seed to start the LSMI (name of the node).

n.wave

Number of waves to be included each the snowball sample

Value

A list of length n.wave where each element is the edge list of edges included on the corresponding wave of sampling.

Details

This is a slightly modified version of the sample_about_one_seed from the snowboot package. This function uses the node names (not node IDs), and it returns the edge list of the bootstrap sample (not the nodes included in the sample).

This is primarily used as an internal function for bootstrap_snowboot. Most code in this function is directly from sample_about_one_seed from the snowboot package. This function is designed to be a wrapper function to format the sample_about_one_seed results into the JaB package syntax.

References

Ramirez-Ramirez L, Nezafati K, Chen Y, Lyubchich Y, Gel YR (2022). snowboot: Bootstrap Methods for Network Inference. R package version 1.0.2, https://cran.r-project.org/web/packages/snowboot.

Gel YR, Lyubchich V, Ramirez LR (2017). “Bootstrap quantification of estimation uncertainties in network degree distributions.” Scientific Reports, 7(1). doi:10.1038/s41598-017-05885-x .

See also