Jackknife After-Bootstrap

Detects influential nodes in a network with the jackknife-after-bootstrap algorithm.

Usage

get_jackknife_after(
  network,
  boot.result,
  quant,
  func.name,
  package.name = NULL,
  func.args = NULL,
  nodes = NULL
)

Arguments

network

An igraph object representing the original network.

boot.result

A list of bootstrap network samples, as produced by a bootstrap_ function.

quant

A numeric value of length 2 specifying the upper quantile used to flag influential nodes (e.g., 0.95).

func.name

A character string specifying the centrality function to be applied (e.g., "degree", "betweenness"). See get_centrality for details.

package.name

(Optional) A character string specifying the package containing func.name. See get_centrality for details.

func.args

(Optional) A list of additional arguments for func.name.See get_centrality for details.

nodes

(Optional) A vector of node names to run JaB algorithm for. If NULL, algorithm will be run on all \(n\) nodes.

Value

A data frame containing:

Node_Number: Numeric IDs of nodes.

Node_Name: Name of nodes.

Orig_Stat: The original centrality statistic of each node.

Upper_Quantile: Upper quantile of the jackknife-after-bootstrap distribution of centrality statistic for each node.

Influential: Logical indicating if each node is influential, i.e. is Orig_Stat greater than Upper_Quantile?

Influential: Logical indicating if there were bootstrap samples that did not include that node, meaning there are jackknife-after sample to generate the distribution of centrality statistics in networks that do not contain that node. If FALSE, then all bootstrap samples contained that node and the Upper_Quantile column will generally be NA. If many nodes are FALSE it could mean that the bootstrap method is poorly tuned and is sampling more nodes that is appropriate for this data set. If only a few nodes are FALSE it could mean the bootstrap method is poorly tuned for this data set, or it could mean that the node is extremely influential as it is highly improbable to generate a bootstrap sample that does not contain that node. Which explanation is appropriate depends on the data set and the bootstrap method used.

Details

This function performs a jackknife after-bootstrap procedure for each node in the network. OUTLINE ALGORITHM

Examples

#add example later